requirements.txt
numpy==2.3.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
speedtest-cli==2.1.3
tzdata==2025.2
